Fixed a concurrency issue
I initially used,

> persistentContainer.performBackgroundTask()

But it displayed really weird issue on the main thread (freezing the table) so I decided to put run the transactions on the main thread instead which was a bad idea (thanks to Jeff!).

So I finally switched to the old school method creating my own private queue context.